pyproject.toml
[project]
name = "mini-context-sphere"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.124.2",
    "pydantic>=2.12.5",
    "uvicorn[standard]>=0.38.0",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "deptry>=0.24.0",
    "mypy>=1.19.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
    "pytest>=7.0.0",
    "httpx>=0.27.0",
    "codecov-cli>=0.6.0",
    "pymarkdownlnt>=0.9.20",
    "mdformat>=0.7.17",
    "mdformat-gfm>=0.3.6",
    "mdformat-frontmatter>=2.0.7",
]

[tool.deptry.per_rule_ignores]
DEP002 = ["uvicorn"]

[tool.ruff]
line-length = 100
target-version = "py314"
src = ["app", "pipeline", "tests"]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle
  "F",  # pyflakes
  "I",  # isort (import sorting)
  "UP", # pyupgrade
  "B",  # flake8-bugbear
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
